+++ src/components/TaskForm.js
@@ -1,6 +1,6 @@
 import { Form } from 'semantic-ui-react'
 import { DateTimeInput } from 'semantic-ui-calendar-react';
-import { useDispatch } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
@@ -10,6 +10,8 @@ const TaskForm = () => {
   const { t } = useTranslation()
 
   const dispatch = useDispatch()
+  const taskReducer = useSelector((state) => state.taskReducer);
+  const { showAddTask } = taskReducer
 
   const [text, setText] = useState('')
   const [day, setDay] = useState('')
@@ -32,7 +34,7 @@ const TaskForm = () => {
       })
   }
 
-  return (
+  return ( showAddTask &&
     <Form onSubmit={onSubmit}>
       <Form.Input required
         label={t('page.AddTask.fields.task.label')}
