+++ src/components/TaskListCard.js
@@ -1,4 +1,4 @@
-import { Button, Card } from 'semantic-ui-react'
+import { Card } from 'semantic-ui-react'
 import { Link } from 'react-router-dom'
 import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
@@ -6,14 +6,16 @@ import { useTranslation } from 'react-i18next'
 
 import TaskCard from './TaskCard'
 import TaskForm from './TaskForm'
-import { fetchTasks, toggleShowAddTask } from '../redux/actions'
+import TaskFormButton from './TaskFormButton'
+
+import { fetchTasks } from '../redux/actions'
 
 const TaskListCard = () => {
   const { t } = useTranslation()
 
   const dispatch = useDispatch()
   const taskReducer = useSelector((state) => state.taskReducer);
-  const { showAddTask, tasks } = taskReducer
+  const { tasks } = taskReducer
 
   useEffect(() => {
     dispatch(fetchTasks())
@@ -22,27 +24,16 @@ const TaskListCard = () => {
       })
   }, [])
 
-  const onToggleShowAddTask = () => {
-    dispatch(toggleShowAddTask(showAddTask))
-      .catch(e => {
-        console.log(e)
-      })
-  }
-
   return (
     <Card fluid>
       <Card.Content>
-        <Button 
-          floated='right' 
-          color={showAddTask ? 'black' : 'blue'}
-          content={showAddTask ? t('page.framework.buttons.close') : t('page.framework.buttons.add')}
-          onClick={() => onToggleShowAddTask()} />
+        <TaskFormButton />
         <Card.Header>
           <h1>{t('page.framework.title')}</h1>
         </Card.Header>
       </Card.Content>
       <Card.Content>
-        {showAddTask && <TaskForm />}
+        <TaskForm />
         <>
         {tasks.length > 0 
           ? (tasks.map((task) => ( <TaskCard key={task.id} task={task} />)))
